moved isguesscorrect into combat deck
